trying to compile

+++ Lab 2/slow.cu
@@ -119,8 +119,8 @@ int main() {
     // pretty_print(size, vec_a, vec_b, mat);
 
 
-    const int32_t numberOfThreadsPerBlock = (int32_t) GRIDSIZE * GRIDSIZE;
-    const int32_t oneDimBlockCount = ceil(size / (double) numberOfThreadsPerBlock);
+    const int numberOfThreadsPerBlock = GRIDSIZE * GRIDSIZE;
+    const int oneDimBlockCount = ceil(size / (double) numberOfThreadsPerBlock);
 
     vectorAdd << oneDimBlockCount, numberOfThreadsPerBlock >> (d_vec_a, d_vec_b, d_out, size);
 
@@ -141,7 +141,7 @@ int main() {
     const int32_t twoDimBlockCount = ceil(size / GRIDSIZE);
     dim3 numBlocks(twoDimBlockCount, twoDimBlockCount, 1);
 
-    matrixMult<<numBlocks, threadsPerBlock>>(size, d_out, d_mat, d_vec_a);
+    matrixMult << numBlocks, threadsPerBlock >> (size, d_out, d_mat, d_vec_a);
 
     cudaError_t cudaerror = cudaDeviceSynchronize(); // waits for completion, returns error code
     if (cudaerror != cudaSuccess)
